+++ Lab 2/trs/Models/GlobalModels.cs
@@ -1,0 +1,10 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace trs.Models;
+
+public class GDataModel
+{
+    public static string Gusername;//global username variable
+    [DataType(DataType.Date)]
+    public static DateTime Gdate;//global date variable
+}