+++ Lab 2/trs/Models/LectureModel.cs
@@ -1,7 +1,0 @@
-namespace trs.Models;
-
-public class LectureModel
-{
-    public string email { get; set; }
-    public string pwd { get; set; }
-}